# projects/doom7/scratch/shared/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Engine-Editor Communication API
  version: "1.0.0"
  description: Minimal internal API for Engine rendering, Level editing contracts, and event streaming hooks between the Engine and Editor components.
servers:
  - url: http://localhost/v1
    description: Local development
paths:
...
  /editor/level:
    get:
      summary: Load the current LevelContract
      operationId: getLevelContract
      responses:
        '200':
          description: Current level contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LevelContract'
              example:
                level_id: "level_01"
                name: "Intro Level"
                version: "v1.0"
                content: { "entities": [], "assets": [] }
    post:
      summary: Save or update a LevelContract
      operationId: saveLevelContract
      requestBody:
        description: LevelContract payload to save
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LevelContract'
      responses:
        '200':
          description: Save result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  level_id:
                    type: string
              example:
                success: true
                level_id: "level_01"
...
components:
  schemas:
...
    LevelContract:
      type: object
      properties:
        level_id:
          type: string
        name:
          type: string
        version:
          type: string
        content:
          type: object
          description: Arbitrary level content (entities, assets, metadata)
          additionalProperties: true
      required:
        - level_id
        - name
        - version
        - content
      example:
        level_id: "level_01"
        name: "Intro Level"
        version: "v1.0"
        content: { "entities": [], "assets": [] }
...